script: Lenovo Legion Go display configuration
Add support for the Lenovo Legion Go handheld, which features a
rotated 1600x2560 panel that reports 60hz and 144hz in the EDID.
VRR and HDR are not supported, and only one panel model is known
to be in use.

Starting with every integer between 60hz and 144hz, valid refresh rates
for the panel were tested one-by-one in two games:

* Hades II
* Portal 2

With dynamic modegen, the range of valid refresh rates can be expanded
to 60hz + 125hz -> 144hz, offering more frame limit notches in the lower
range of the Steam UI unified slider. This new range is what was used
for further validation across several units.

Linux setups tested include:

* Bazzite 40 with kernel 6.9.12-210.fsync
* SteamOS Main 20240919.1002 with kernel 6.8.12-valve2-1

Games tested with the new range include:

* Hades II
* Portal 2
* Left 4 Dead 2
* Metro Exodus
* Outer Wilds
* God of War: Ragnarok
* Hi-Fi Rush
* Manor Lords
* Nine Sols

All games tested were able to run with the slider behaving like SteamOS
on a Steam Deck LCD when modesetting and maintaining a frame limit.

Tested-by: Kyle Gospodnetich <kylegospo@bazzite.gg>
Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>